tests/functional/Bots/Python/Consumers/CodeFirst/WaterfallHostBot/requirements.txt
botbuilder-integration-aiohttp>=4.13.0
botbuilder-dialogs>=4.13.0
python-dotenv~=0.15.0
